Corrected a problem that the internal variable of '_ldap_uri' might be changed, and may cause a LDAP connection error.
When a Referral LDAP Directory is loaded, the internal variable of
'_ldap_uri' is updated to re-connect it.
But this value isn't saving and returning before and after calling the
'_get_ldap_search_referral' method (Sorry, this is a bug that I made).
By this bug, that internal variable might be changed after searching the
LDAP tree, then it may cause LDAP connection error.

This commit correct it.